Move reusable functions into helper folders
