+++ frontend/common/src/helpers/skillUtils.ts
@@ -130,4 +130,12 @@ export function filterSkillsByNameOrKeywords(
   return matchedSkills;
 }
 
+export const getMissingSkills = (required: Skill[], added?: Skill[]) => {
+  return !added?.length
+    ? required
+    : required.filter((skill) => {
+        return !added.find((addedSkill) => addedSkill.id === skill.id);
+      });
+};
+
 export default { invertSkillSkillFamilyTree, invertSkillExperienceTree };
